Add static_cast where needed for assertEqual
In some places assertEqual was called with (size_t, int)
which led to compiler warnings. Fix those with
static_cast<std::size_t> of the int argument.
In one case the size_t argument was cast to int instead,
but I changed it to match the rest of the code.

+++ include/astshim/Frame.h
@@ -1125,7 +1125,7 @@ public:
     axis order for output data).
     */
     void permAxes(std::vector<int> perm) {
-        detail::assertEqual(static_cast<int>(perm.size()), "perm.size()", getNaxes(), "naxes");
+        detail::assertEqual(perm.size(), "perm.size()", static_cast<std::size_t>(getNaxes()), "naxes");
         astPermAxes(getRawPtr(), perm.data());
         assertOK();
     }
